Give demo input a default value.
I was confused by the demo at first, because I saw the placeholder,
and thought it was an actual value. I hit the "equals" button
and got a parse error (doh!). By giving the demo input a default 
value, a user can just hit "equals" and get a result right away.