add CONTRIBUTING.md with guidelines
It was a bit confusing for me when I started, because I
wasn't familiar with Makefiles, so I added some details here.

I like the CONTRIBUTING.md file, because Github pops up a little
message when you open a PR on the project, to advise you
to read it. Let me know if this suits your tastes. :)