BUGFIX: fix collision in manual aliases and auto-aliases when user assigns an alias to a term while other terms in the same rule get auto-aliased: the latter didn't check if these autogenerated aliases would collide with user-assigned aliases! Rule example: `e: WORD[e1] + e` would collide auto-alias `e1` created for the `e` term with the user-assigned `e1` alias for the `WORD` term!
---

When the rule is fitted with aliases it doesn't mean that the action code MUST use those:
we therefor allow access to both the original (non)terminal and the alias.

Also note that each (non)terminal can also be uniquely addressed by [$@]<nonterminal><N>
where N is a number representing the number of this particular occurrence of the given
(non)terminal.

For example, given this (intentionally contrived) production:

     elem[alias] elem[another_alias] another_elem[alias] elem[alias] another_elem[another_alias]

all the items can be accessed as:

     $1 $2 $3 $4 $5
     $elem1 $elem2 $another_elem1 $elem3 $another_elem2
     $elem $elem2 $another_elem $elem3 $another_elem2
     $alias1 $another_alias1 $alias2 $alias3 $another_alias2
     $alias $another_alias $alias2 $alias3 $another_alias2

where each line above is equivalent to the top-most line. Note the numbers postfixed to
both (non)terminal identifiers and aliases alike and also note alias2 === another_elem1:
the postfix numbering is independent.

WARNING: this feature is disabled for a term when there already exists an
(human-defined) *alias* for this term *or* when the numbered auto-alias already
exists because the user has used it as an alias for another term, e.g.

     e: WORD[e1] '=' e '+' e;

would *not* produce the `e1` and `e2` aliases, as `e1` is already defined
as an explicit alias: adding auto-alias `e1` would then break the system,
while `e2` would be ambiguous from the human perspective as he *might* then
expect `e2` and `e3`.