BUGFIX: stop parse kernel from running ad infinitum when an error is caught and handled just before we hit EOF, while the parse stack still contains further parse error(s) in the parent rules.
We need to make sure we're not cycling forever:
once we hit EOF, even when we `yyerrok()` an error, we must
prevent the core from running forever,
e.g. when parent rules are still expecting certain input to
follow after this, for example when you handle an error inside a set
of braces which are matched by a parent rule in your grammar.

Hence we require that every error handling/recovery attempt
*after we've hit EOF* has a diminishing state stack: this means
we will ultimately have unwound the state stack entirely and thus
terminate the parse in a controlled fashion even when we have
very complex error/recovery code interplay in the core + user
action code blocks.