fiddling with the error recovery code:
Also check if the LookAhead symbol isn't the ERROR token we set as part of the error
recovery, for then this we would we idling (cycling) on the error forever.
Yes, this does not take into account the possibility that the *lexer* may have
produced a *new* TERROR token all by itself, but that would be a very peculiar grammar!